requirements.txt
fastapi
uvicorn[standard]
seleniumbase
beautifulsoup4
colorama
pyTelegramBotAPI
requests
motor
pymongo
pyyaml